+++ netring/src/prelude.rs
@@ -29,8 +29,12 @@ pub use crate::monitor::HttpFingerprint;
 pub use crate::monitor::TlsFingerprint;
 #[cfg(feature = "dhcp")]
 pub use crate::protocol::builtin::Dhcp;
+#[cfg(feature = "dnp3")]
+pub use crate::protocol::builtin::Dnp3;
 #[cfg(feature = "dns")]
 pub use crate::protocol::builtin::Dns;
+#[cfg(feature = "ftp")]
+pub use crate::protocol::builtin::Ftp;
 #[cfg(feature = "http")]
 pub use crate::protocol::builtin::Http;
 pub use crate::protocol::builtin::Icmp;
@@ -38,6 +42,8 @@ pub use crate::protocol::builtin::Icmp;
 pub use crate::protocol::builtin::Kerberos;
 #[cfg(feature = "ldap")]
 pub use crate::protocol::builtin::Ldap;
+#[cfg(feature = "modbus")]
+pub use crate::protocol::builtin::Modbus;
 #[cfg(feature = "netbios-ns")]
 pub use crate::protocol::builtin::Nbns;
 #[cfg(feature = "ntp")]
@@ -50,14 +56,20 @@ pub use crate::protocol::builtin::Radius;
 pub use crate::protocol::builtin::Rdp;
 #[cfg(feature = "smb")]
 pub use crate::protocol::builtin::Smb;
+#[cfg(feature = "smtp")]
+pub use crate::protocol::builtin::Smtp;
 #[cfg(feature = "snmp")]
 pub use crate::protocol::builtin::Snmp;
 #[cfg(feature = "ssdp")]
 pub use crate::protocol::builtin::Ssdp;
 #[cfg(feature = "ssh")]
 pub use crate::protocol::builtin::Ssh;
+#[cfg(feature = "stun")]
+pub use crate::protocol::builtin::Stun;
 #[cfg(feature = "tftp")]
 pub use crate::protocol::builtin::Tftp;
+#[cfg(feature = "wireguard")]
+pub use crate::protocol::builtin::WireGuard;
 pub use crate::protocol::builtin::{Tcp, Udp};
 #[cfg(feature = "tls")]
 pub use crate::protocol::builtin::{Tls, TlsHandshake};